requirements.txt
python-docx==0.8.11
PyQt6==6.4.2
tqdm==4.65.0